Move navigation content based service to Sulu Content Infrastructure integration
