+++ packages/page/tests/Unit/Infrastructure/Sulu/Content/DataMapper/NavigationContextDataMapperTest.php
@@ -9,13 +9,13 @@
  * with this source code in the file LICENSE.
  */
 
-namespace Sulu\Bundle\Page\Tests\Unit\Application\Merger\DataMapper;
+namespace Sulu\Bundle\Page\Tests\Unit\Infrastructure\Sulu\Content\DataMapper;
 
 use PHPUnit\Framework\TestCase;
 use Prophecy\PhpUnit\ProphecyTrait;
-use Sulu\Page\Application\DataMapper\NavigationContextDataMapper;
 use Sulu\Page\Domain\Model\Page;
 use Sulu\Page\Domain\Model\PageDimensionContent;
+use Sulu\Page\Infrastructure\Sulu\Content\DataMapper\NavigationContextDataMapper;
 
 class NavigationContextDataMapperTest extends TestCase
 {
